requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pypdf
scikit-learn